% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtClimate}
\alias{dtClimate}
\title{Climate data.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 23628 rows and 5 columns.}
\source{
http://doi.org/10.7289/V5D21VHZ
Downloaded from ftp://ftp.ncdc.noaa.gov/pub/data/ghcn/daily/by_year/

Menne, M.J., I. Durre, B. Korzeniewski, S. McNeal, K. Thomas, X. Yin,
S. Anthony, R. Ray, R.S. Vose, B.E.Gleason, and T.G. Houston,
2012: Global Historical Climatology Network - Daily, Version 3.12
}
\usage{
data(dtClimate)
}
\description{
The climate data used in the blogpost.
}
\keyword{datasets}
