% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_marimekko.R
\name{Marimekkofy}
\alias{Marimekkofy}
\title{Marimekkofy}
\usage{
Marimekkofy(data, xbucket = "xbucket", ybucket = "ybucket",
  weight = NULL)
}
\arguments{
\item{data}{dataframe}

\item{xbucket}{x value}

\item{ybucket}{y value}

\item{weight}{weight value}
}
\value{
df
}
\description{
Marimekkofy
}
