% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_calendar_heatmap.R
\name{stat_calendar_heatmap}
\alias{stat_calendar_heatmap}
\title{Plots a calendar heatmap}
\usage{
stat_calendar_heatmap(mapping = NULL, data = NULL, show.legend = NA,
  inherit.aes = TRUE, na.rm = T, bandwidth = NULL, ...)
}
\description{
A calendar heatmap provides context for weeks, and day of week and
is a better way to visualise daily data than line charts.
}
\section{Aesthetics}{
 date, fill.
}

\section{Data Tips}{
 \code{\link{base::strftime}} can help extract the value
of the year, week of year, and day of week from the date column. You might
need to extract the year to facet multiple years as demonstrated in the
example. \cr\cr
This stat uses the following transformation to obtain the x and  y
coordinate to be used in the heatmap -
\code{ \cr
   data$x = 1 + as.integer(strftime(data$date, "\%W")) \cr
   data$y = as.integer(strftime(data$date, "\%w")) \cr
   data$y[data$y == 0L] = 7 \cr
   data$y = 8 - data$y \cr
}
}

\section{Cosmetic Tips}{
 The minimalist look can be achieved by appending the
following chunk of code to the output object:
\code{ \cr
+ \cr
xlab(NULL) + \cr
ylab(NULL) + \cr
scale_fill_continuous(low = 'green', high = 'red') + \cr
theme( \cr
   axis.text = element_blank(), \cr
   axis.ticks = element_blank(), \cr
   legend.position = 'none', \cr
   strip.background = element_blank(), \cr
   panel.background = element_blank(), \cr
   panel.border = element_blank() \cr
) \cr
}
}

\section{Also See}{
 \code{\link{ggplot_calendar_heatmap}}, a
polished but less flexible alternative.
}

\examples{
DailyData = data.frame(
      DateCol = seq(
         as.Date("1/01/2014", "\%d/\%m/\%Y"),
         as.Date("31/12/2015", "\%d/\%m/\%Y"),
         "days"
      ),
      ValueCol = runif(730)
   )
   DailyData$Year = strftime(DailyData$DateCol, "\%Y")
ggplot(
   DailyData,
   aes(
      date = DateCol,
      fill = ValueCol
   )
) +
   stat_calendar_heatmap() +
   facet_wrap(~Year, ncol = 1)
}
