% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_waterfall.R
\name{stat_waterfall}
\alias{stat_waterfall}
\title{Plot a time series as a waterfall plot}
\usage{
stat_waterfall(mapping = NULL, data = NULL, show.legend = NA,
  inherit.aes = TRUE, na.rm = T, bandwidth = NULL, ...)
}
\description{
A waterfall plot highlights the change in the time series rather than the
value of the time series itself.
}
\section{Aesthetics}{
 stat_waterfall needs x, y
}

\section{Cosmetic Tips}{
 The minimalist look can be achieved by appending the
following chunk of code to the output object:
\code{ \cr
+ \cr
xlab(NULL) + \cr
ylab(NULL) + \cr
theme( \cr
   axis.text = element_blank(), \cr
   axis.ticks = element_blank(), \cr
   legend.position = 'none', \cr
   strip.background = element_blank(), \cr
   plot.background = element_blank(), \cr
   panel.background = element_blank(), \cr
   panel.border = element_blank(), \cr
   panel.grid = element_blank(), \cr
   panel.border = element_blank() \cr
) \cr
}
}

\section{Also See}{
 \code{\link{ggplot_waterfall}}, a
flexible but less polished alternative.
}

\examples{
set.seed(1)
dfData = data.frame(x = 1:20, y = cumsum(rnorm(20)))
ggplot(dfData, aes(x =x, y = y) )+
   stat_waterfall()
}
