% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_calendar_heatmap.R
\docType{data}
\name{StatCalendarMonthBreaks}
\alias{StatCalendarMonthBreaks}
\title{Transforms data for the month breaks of the calendar heatmap}
\format{An object of class \code{StatCalendarMonthBreaks} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.}
\usage{
StatCalendarMonthBreaks
}
\description{
Transforms data for the month breaks of the calendar heatmap
}
\keyword{datasets}
