% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_calendar_heatmap.R
\docType{data}
\name{StatCalendarTiles}
\alias{StatCalendarTiles}
\title{Transforms data for the tiles of the heatmap}
\format{An object of class \code{StatCalendarTiles} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\usage{
StatCalendarTiles
}
\description{
Transforms data for the tiles of the heatmap
}
\keyword{datasets}

