% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_horizon.R
\docType{data}
\name{StatHorizon}
\alias{StatHorizon}
\title{Transforms data for a horizon plot}
\format{An object of class \code{StatHorizon} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
StatHorizon
}
\description{
Transforms data for a horizon plot
}
\keyword{datasets}
