% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_marimekko.R
\docType{data}
\name{StatMarimekko}
\alias{StatMarimekko}
\title{Transforms data for the tiles of the heatmap}
\format{An object of class \code{StatMarimekko} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
StatMarimekko
}
\description{
Transforms data for the tiles of the heatmap
}
\keyword{datasets}
