% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_occurrence.R
\docType{data}
\name{StatOccurrence}
\alias{StatOccurrence}
\title{Transforms data for an occurrence plot}
\format{An object of class \code{StatOccurrence} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.}
\usage{
StatOccurrence
}
\description{
Transforms data for an occurrence plot
}
\keyword{datasets}
