% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_steamgraph.R
\docType{data}
\name{StatSteamgraph}
\alias{StatSteamgraph}
\title{Transforms data for a steam graph}
\format{An object of class \code{StatSteamgraph} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
StatSteamgraph
}
\description{
Transforms data for a steam graph
}
\keyword{datasets}
