% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_waterfall.R
\docType{data}
\name{StatWaterfall}
\alias{StatWaterfall}
\title{Transforms data for a horizon plot}
\format{An object of class \code{StatWaterfall} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.}
\usage{
StatWaterfall
}
\description{
Transforms data for a horizon plot
}
\keyword{datasets}
